+++ lib/animation.dart
@@ -1,0 +1,3 @@
+library flutter_eval.animation;
+
+export 'src/animation/curves.dart' show $Curve;