+++ lib/painting.dart
@@ -1,0 +1,3 @@
+library flutter_eval.painting;
+
+export 'src/painting/basic_types.dart' show $Axis;