+++ lib/src/animation.dart
@@ -1,0 +1,5 @@
+const animationSource = '''
+library animation;
+
+export 'package:flutter/src/animation/curves.dart';
+''';